+++ ruoyi-ee/src/main/java/com/ruoyi/ee/domain/CwOutWork.java
@@ -1,5 +1,6 @@
 package com.ruoyi.ee.domain;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -26,11 +27,11 @@ public class CwOutWork extends BaseEntity
 
     /** 总天数 */
     @Excel(name = "总天数")
-    private Long allDays;
+    private Double allDays;
 
     /** 总时数 */
     @Excel(name = "总时数")
-    private Long allHours;
+    private Double allHours;
 
     /** 申请状态 */
     @Excel(name = "申请状态")
@@ -83,21 +84,21 @@ public class CwOutWork extends BaseEntity
     {
         return onoffStationId;
     }
-    public void setAllDays(Long allDays) 
+    public void setAllDays(Double allDays)
     {
         this.allDays = allDays;
     }
 
-    public Long getAllDays() 
+    public Double getAllDays()
     {
         return allDays;
     }
-    public void setAllHours(Long allHours) 
+    public void setAllHours(Double allHours)
     {
         this.allHours = allHours;
     }
 
-    public Long getAllHours() 
+    public Double getAllHours()
     {
         return allHours;
     }
